fix: Check duplicates of both src and dest
This commit moves code to check duplicates from `pull_a_chain()` to
`sort_actions()` and made it simpler. Because of the simplification,
there is no need to distinguish whether duplicates happens at a first
action of a chain or not. Hence, now we have a test case for source
duplication and for destination duplications.